+++ examples/noise_handshake.rs
@@ -6,7 +6,7 @@ use tracing_subscriber::filter::{EnvFilter, LevelFilter};
 
 use pea2pea::{
     connections::ConnectionSide,
-    protocols::{Handshaking, Reading, Writing},
+    protocols::{Handshaking, Reading, ReturnableConnection, Writing},
     Node, NodeConfig, Pea2Pea,
 };
 
@@ -83,8 +83,7 @@ impl SecureNode {
 
 impl Handshaking for SecureNode {
     fn enable_handshaking(&self) {
-        let (from_node_sender, mut from_node_receiver) = mpsc::channel(1);
-        self.node().set_handshake_handler(from_node_sender.into());
+        let (from_node_sender, mut from_node_receiver) = mpsc::channel::<ReturnableConnection>(1);
 
         // the noise handshake settings used by snow
         const HANDSHAKE_PATTERN: &str = "Noise_XXpsk3_25519_ChaChaPoly_BLAKE2s";
@@ -92,7 +91,7 @@ impl Handshaking for SecureNode {
 
         // spawn a background task dedicated to handling the handshakes
         let self_clone = self.clone();
-        tokio::spawn(async move {
+        let handshaking_task = tokio::spawn(async move {
             loop {
                 if let Some((mut conn, result_sender)) = from_node_receiver.recv().await {
                     let noise_state = match !conn.side {
@@ -188,6 +187,9 @@ impl Handshaking for SecureNode {
                 }
             }
         });
+
+        self.node()
+            .set_handshake_handler((from_node_sender, handshaking_task).into());
     }
 }